HS-34-02: split web/routes/activity.py into routes/activity/ sub-package
The 1,319-line / 38-handler build_activity_router(ctx) becomes a routes/activity/
sub-package by domain, behind a STABLE build_activity_router(ctx) (routes/
__init__.py unchanged):

- ledger.py       /api/activity/status, records, refresh, settings, domains
- rules.py        /api/activity/project-rules*
- enrichment.py   /api/activity/enrichment/*, extension/events, annotations, briefing
- candidates.py   /api/activity/meeting-candidates*
- plugin_jobs.py  /api/plugin-jobs*

Behavior-preserving: handlers moved verbatim (only +1 relative-import dot);
absolute paths + include_router keep the route table identical (38 routes, hash
d4332051064ff059 unchanged). Simpler than the dictation split — every module-level
payload shaper is used by exactly one group, so each travels with its handlers and
there is no shared state to thread; the only cross-cutting helpers still come from
the neutral web/runtime_support.

New tests/unit/test_activity_routes_split.py locks the exact (path, method) set.
Suite green 1956/15; package ruff + F821 clean.
